Remove BasePackagePermissionsAPIView and refactor
* Remove the BasePackagePermissionsAPIView which
  PackagePermissionsAPIView was inheriting as it was redundant.
* Move serializers from view fiel to serializers file.
* Remove empty file test_detail_view_permissions.py.

Refs. TS-2284
