+++ django/thunderstore/api/cyberstorm/views/package_permissions.py
@@ -1,42 +1,20 @@
 from django.http import HttpRequest
-from rest_framework import serializers, status
+from rest_framework import status
 from rest_framework.generics import get_object_or_404
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from thunderstore.api.cyberstorm.serializers import PackagePermissionsSerializer
 from thunderstore.api.utils import conditional_swagger_auto_schema
 from thunderstore.repository.models import Community, PackageListing
 from thunderstore.repository.views.package._utils import get_package_listing_or_404
 from thunderstore.repository.views.package.detail import PermissionsChecker
 
 
-class PackageInfoSerializer(serializers.Serializer):
-    community_id = serializers.CharField()
-    namespace_id = serializers.CharField()
-    package_name = serializers.CharField()
-
-
-class PermissionsSerializer(serializers.Serializer):
-    can_manage = serializers.BooleanField()
-    can_manage_deprecation = serializers.BooleanField()
-    can_manage_categories = serializers.BooleanField()
-    can_deprecate = serializers.BooleanField()
-    can_undeprecate = serializers.BooleanField()
-    can_unlist = serializers.BooleanField()
-    can_moderate = serializers.BooleanField()
-    can_view_package_admin_page = serializers.BooleanField()
-    can_view_listing_admin_page = serializers.BooleanField()
-
-
-class PackagePermissionsSerializer(serializers.Serializer):
-    package = PackageInfoSerializer()
-    permissions = PermissionsSerializer()
-
-
-class BasePackagePermissionsAPIView(APIView):
+class PackagePermissionsAPIView(APIView):
     """
-    Base class for getting the permissions for the current user on a package listing.
+    View for getting the permissions for the current user on a package.
     """
 
     permission_classes = [IsAuthenticated]
@@ -67,12 +45,6 @@ class BasePackagePermissionsAPIView(APIView):
 
         permission_data = permissions_checker.get_permissions()
         return permission_data
-
-
-class PackagePermissionsAPIView(BasePackagePermissionsAPIView):
-    """
-    View for getting the permissions for the current user on a package.
-    """
 
     @conditional_swagger_auto_schema(
         responses={200: PackagePermissionsSerializer},
